requirements.txt
pandas==2.3.3
plotly==6.3.1
python-dotenv==1.1.1
snowflake_connector_python==4.0.0
streamlit==1.50.0
